Pulls out rollup-specific common.
